+++ src/app/(characters)/api/get-character-by-id.ts
