+++ src/app/(characters)/characters/[characterId]/components/no-comics-found.tsx
@@ -1,0 +1,3 @@
+export const NoComicsFound = () => {
+  return <p className="text-gray-600">No comics found.</p>
+}
