--- src/components/ui/pagination/index.ts
@@ -1,1 +1,0 @@
-export * from './pagination'
